tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/readyState/001.html

<!doctype html>
<title>WebSockets: getting readyState in connecting</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function(t) {
  assert_equals((new WebSocket(SCHEME_DOMAIN_PORT+'/')).readyState, WebSocket.CONNECTING);
});
</script>
